requirements.txt
# apex==0.1
appdirs==1.4.4
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cycler==0.10.0
Cython==0.29.21
DateTime==4.3
decorator==4.4.2
deepdish==0.3.6
defusedxml==0.6.0
distlib==0.3.1
dlib==19.17.0
dominate==2.5.1
easydict==1.9
entrypoints==0.3
fasteners==0.15
filelock==3.0.12
fire==0.3.1
fjcommon==0.2.12
future==0.18.2
h5py==2.10.0
hypertools==0.6.2
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
importlib-resources==3.0.0
ipdb==0.13.3
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.6
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==2.2.0
jupyterlab-server==1.2.0
kiwisolver==1.2.0
leveldb==0.201
llvmlite==0.33.0
lmdb==0.98
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
monotonic==1.5
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.4
nose==1.3.7
notebook==6.0.3
numba==0.50.1
numexpr==2.7.1
numpy==1.19.0
opencv-python==4.3.0.36
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
# panopticapi #@ git+https://github.com/cocodataset/panopticapi.git@118ffaba478b5d45925e88d504625b4123f3160a
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
ppca==0.0.4
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
pycocotools #@ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
pycparser==2.20
# pycurl #==7.43.0
pydicom==2.0.0
Pygments==2.6.1
# pygobject==3.20.0
pyparsing==2.4.7
pyrsistent==0.16.0
# python-apt==1.1.0b1+ubuntu0.16.4.9
python-dateutil==2.8.1
python-gflags==3.1.2
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
requests==2.24.0
scikit-image==0.17.2
scikit-learn #==0.23.1
scikit-video==1.1.11
scipy==1.5.1
seaborn==0.10.1
Send2Trash==1.5.0
SimpleITK==1.2.4
six==1.15.0
sklearn==0.0
tables==3.6.1
tensorboardX==2.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.7.17
torchfile==0.1.0
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
urllib3==1.25.9
virtualenv==20.0.27
visdom==0.1.8.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
yacs==0.1.7
zipp==3.1.0
zope.interface==5.1.0
pyyaml 
tqdm 
tensorboardX 
anyconfig 
munch 
scipy 
sortedcontainers 
shapely 
pyclipper 
gevent 
gevent-websocket 
flask 
editdistance 
scikit-image 
imgaug==0.2.8